requirements.txt
# FastMCP framework for building MCP servers
fastmcp>=0.5.2

# ASGI server for running the FastMCP application
uvicorn[standard]>=0.30.0

# Looker Python SDK
looker-sdk>=24.0.0

# Web framework and HTTP client
starlette>=0.37.0
httpx>=0.27.0
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0
